eco-bee src/backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6
Pillow>=10.0.0
transformers>=4.35.0
torch>=2.0.0
numpy>=1.24.0
mistralai>=0.1.0
requests>=2.31.0
python-dotenv>=1.0.0
accelerate>=0.24.0
datasets>=2.14.0
scikit-learn>=1.3.0
pandas>=2.0.0
sqlalchemy>=2.0.0
alembic>=1.12.0
supabase>=1.0.0